Very rough first draft at codegen module
